fix: resolve CodeQL go/incorrect-integer-conversion
Avoid unsafe narrowing of int64 to int32 in hosts.go memory calculation
which could overflow for hosts with >2GB RAM. Add bounds validation for
pageSize before int32 conversion in eventDispacher.go.

Resolves: NR-560201

+++ internal/process/hosts.go
@@ -110,7 +110,7 @@ func createHostSamples(config *config.Config) {
 			memoryUsed := host.Summary.QuickStats.OverallMemoryUsage
 			checkError(config.Logrus, ms.SetMetric("mem.usage", memoryUsed, metric.GAUGE))
 
-			memoryFree := int32(memoryTotal) - memoryUsed
+			memoryFree := memoryTotal - int64(memoryUsed)
 			checkError(config.Logrus, ms.SetMetric("mem.free", memoryFree, metric.GAUGE))
 
 			// cpu
